requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.9.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==14.2.0
mysqlclient==2.1.1
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
